+++ components/files/odv_file.h
@@ -20,8 +20,8 @@ struct ODVFile
 {
     char filename[FILENAME_MAX];
     char *buf;
-    unsigned int length;
-    unsigned int pos;
+    size_t length;
+    size_t pos;
 
 #ifdef WINDOWS
     HANDLE file;
@@ -34,8 +34,8 @@ struct ODVFile
 struct ODVFile *odv_file_open(const char *filename);
 void odv_file_info(const struct ODVFile *file);
 
-int odv_file_read(struct ODVFile *file, void *buf, size_t count);
-int odv_file_readline(struct ODVFile *file, char *buf, size_t count);
+size_t odv_file_read(struct ODVFile *file, void *buf, size_t count);
+size_t odv_file_readline(struct ODVFile *file, char *buf, size_t count);
 void odv_file_seek(struct ODVFile *file, unsigned int offset);
 int odv_file_close(struct ODVFile *file);
